+++ lib/github/apps/installations.ex
@@ -1,5 +1,5 @@
 defmodule Github.Apps.Installations do
-  import Github
+  import Github.Client
 
   @list_repos_default_options %{
     page: 1,
@@ -10,7 +10,7 @@ defmodule Github.Apps.Installations do
   ## Example
 
       iex> %Github.Client{access_token: "access_token"} |> Github.Apps.Installations.list_repos!(page: 1, per_page: 30)
-      %Github.Response{
+      %Github.Client.Response{
         body: [%{"email" => "email@example.com", "primary" => true, "verified" => true, "visibility" => "private"}],
         status: 200,
         headers: [...]
@@ -24,6 +24,6 @@ defmodule Github.Apps.Installations do
         {"Authorization", "token #{github_client.access_token}"},
         {"Accept", "application/vnd.github.machine-man-preview+json"}
       ]
-    ) |> to_github_response(github_client)
+    )
   end
 end