+++ lib/github/users/emails.ex
@@ -1,5 +1,5 @@
 defmodule Github.Users.Emails do
-  import Github
+  import Github.Client
 
   @list_default_options %{
     page: 1,
@@ -10,7 +10,7 @@ defmodule Github.Users.Emails do
   ## Example
 
       iex> %Github.Client{access_token: "access_token"} |> Github.Users.Emails.list!(page: 1, per_page: 30)
-      %Github.Response{
+      %Github.Client.Response{
         body: [%{"email" => "email@example.com", "primary" => true, "verified" => true, "visibility" => "private"}],
         status: 200,
         headers: [...]
@@ -22,6 +22,6 @@ defmodule Github.Users.Emails do
     get!(
       "https://api.github.com/user/emails?page=#{opts.page}&per_page=#{opts.per_page}",
       [{"Authorization", "token #{github_client.access_token}"}]
-    ) |> to_github_response(github_client)
+    )
   end
 end