+++ CHANGELOG.md
@@ -21,7 +21,7 @@ that you can set version constraints properly.
 #### [v0.4.0](https://github.com/WorkflowCI/github/compare/v0.3.0...v0.4.0) – 2018-08-31
 
 * `Added`: `Github.Orgs.find!`.
-* `Added`: `Github.Response`.
+* `Added`: `Github.Client.Response`.
 
 #### [v0.3.0](https://github.com/WorkflowCI/github/compare/v0.2.0...v0.3.0) – 2018-08-30